4 - выполнил рефакторинг компонента Game
1) классовый компонент Game переписал на функциональный компонент для улучшения простоты читаемости кода.
2) компонент Game.js перенёс в отдельную папку для того, чтобы в дальнейшем вынести фрагменты AppBar и TabContainer в отдельный компонент вместе со стилями для более лучшей читаемости кода.
