+++ src/AC.js
@@ -322,16 +322,16 @@ export const updateGame = (game) => {
                     type: UPLOAD_GAME + SUCCESS,
                     payload: {id: game.id},
                 });
-                dispatch({
-                    type: LOAD_GAMES + SHOULD_RELOAD,
-                });
+                // dispatch({
+                //     type: LOAD_GAMES + SHOULD_RELOAD,
+                // });
 
             })
             .catch((err) => {
-                dispatch({
-                    type: UPLOAD_GAME + FAIL,
-                    payload: { id: game.id, error: err.message },
-                });
+                // dispatch({
+                //     type: UPLOAD_GAME + FAIL,
+                //     payload: { id: game.id, error: err.message },
+                // });
                 console.error(`Не получилось обновить таблицу игры! ID игры: ${game.id}`, err);
             })
     }
@@ -373,10 +373,10 @@ export const updateLog = (log) => {
 
             })
             .catch((err) => {
-                dispatch({
-                    type: UPLOAD_LOG + FAIL,
-                    payload: { message: err.message, id: log.id },
-                });
+                // dispatch({
+                //     type: UPLOAD_LOG + FAIL,
+                //     payload: { message: err.message, id: log.id },
+                // });
                 console.error(`Не получилось обновить таблицу лога! ID игры: ${log.id}`, err);
             })
     }
